+++ weather@mockturtl/src/3_8/providers/open-meteo/payload/common.ts
@@ -14,8 +14,8 @@ export function OpenMeteoWeatherCodeToCondition(code: number, isDay: boolean): C
 			return {
 				icons: !isDay ? ["weather-few-clouds-night"] : ["weather-few-clouds"],
 				customIcon: !isDay ? "night-alt-cloudy-symbolic" : "day-cloudy-symbolic",
-				main: _("Mainly clear"),
-				description: _("Mainly clear")
+				main: _("Few clouds"),
+				description: _("Few clouds")
 			}
 		case 2:
 			return {